added MLX, player handling, raycasting mechanics, rendering and collision detection
For Collision Detection: Implemented a simple Axis-Aligned Bounding Box

About the Raycasting Mechanics: Introduced the concept of Field of View (FOV) and the distance to the projection plane. Implemented the process of casting individual rays for each vertical slice of the screen to determine how far they travel before hitting a wall. Derived the height of each wall slice on the screen based on the distance the ray traveled. Textured the wall slices based on the orientation of the wall they hit.